requirements.txt
# Job Application Assistant Dependencies
# Core
python-jobspy>=0.46.0
pandas>=2.0.0
requests>=2.31.0

# LLM Integration
openai>=1.0.0
anthropic>=0.18.0

# Slack Integration
slack-sdk>=3.27.0
slack-bolt>=1.18.0

# Browser Automation (for auto-apply)
playwright>=1.40.0

# Document Processing
python-docx>=1.1.0
PyPDF2>=3.0.0
reportlab>=4.0.0

# Data Storage
aiofiles>=23.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
schedule>=1.2.0
